LayoutTests/imported/w3c/web-platform-tests/resource-timing/TAO-port-mismatch-means-crossorigin-expected.txt

Description

This test validates that for a cross origin resource with different ports, the timing allow check algorithm will fail when the value of Timing-Allow-Origin value has the right host but the wrong port in it.


PASS A port mismatch must fail the TAO check
FAIL An identical port must pass the TAO check assert_greater_than: transferSize should be greater than 0 expected a number greater than 0 but got 0
